OpenWrt: bump to v23.05 from 2024-03-08
OpenWrt v23.05 releases up to and including v23.05.2 contain a bug
which causes some SPI flashes to be partially or fully unwriteable [1].

A workaround for this bug has already been added to the v23.05 branch,
but no new version has been released since.

Bump OpenWrt and corresponding feeds to the most recent commit
on the v23.05 branch.

[1] https://github.com/openwrt/openwrt/pull/14361

Signed-off-by: Fabian Bläse <fabian@blaese.de>