Add support for Xiaomi Mi Router 4A (100m-intl)
The international variant of the Xiaomi Mi Router 4A (100m) has a
different partition layout as the chinese version and was added to
OpenWrt at a later time. Using the OpenWrt image for the international
variant saves the extra step of flashing the chinese firmware variant
via TFTP before OpenWrt itself.

Signed-off-by: Fabian Bläse <fabian@blaese.de>