fff-dhcp: do not use dnsmasq for resolving locally
A change to the OpenWrt dnsmasq packages introduced with openwrt-24.10
invertes the default state of the localuse option.

However, our firmware actually expects the old behavior, because it is
necessary to resolve hostnames before the configured dnsmasq servers are
reachable (i.e. to resolve tunnel endpoints).

Explicitly set localuse to off, so the dnsmasq resolver is only used by
clients, but not the router itself.

Signed-off-by: Fabian Bläse <fabian@blaese.de>