ntpd: disable ujail to work around bind mount limitations
By default, ntpd is executed in a sandbox. ujail uses bind mounts to
give ntpd access to specific files, including resolv.conf.

/etc/resolv.conf is a symlink which ultimately points to
/tmp/resolv.conf.d/resolv.conf.auto. The bind mount is created on a
single file, therefore the mount breaks as soon as the file gets
deleted. The bind mount still refers to the deleted source file, even
if a new file is created in the same location.

The problem can be avoided by not running ujail in a sandbox. The
sandbox is disabled by removing the ntpd capabilities file.

Upstream issues:
- https://github.com/openwrt/openwrt/issues/10389
- https://github.com/openwrt/openwrt/issues/10843

Fixes: #291

Signed-off-by: Fabian Bläse <fabian@blaese.de>