build_patches: disable new OpenWrt upgrade mechanism
Starting with openwrt-24.10, OpenWrt introduces a new upgrade mechanism
to keep the enabled/disabled state of services across firmware updates.
Changing the state of services is neither expected, nor supported in the
Freifunk Franken firmware.

Remove the the upgrade mechanism to prevent services to be in an
unexpected state after a firmware update.

Signed-off-by: Fabian Bläse <fabian@blaese.de>