mvebu: rb5009: prepare migration for upstream partition layout
The Mikrotik rb5009 has been merged upstream. The partition layout was
changed compared to the development state we have included in our
firmware. This requires migration steps during the update.

Instead of patching the vendors bootloader to make it load fit images, a
u-boot in ELF format is placed into a small YAFFS partition. This way,
the original bootloader (but also a previously patched one) is able to
load the u-boot. Later on, u-boot can load a regular OpenWrt FIT image.

However, the YAFFS partition has to be placed at the beginning of the
NAND flash, which is currently completely occupied by the ubifs.
Therefore, the ubifs partition has been moved.

To migrate from the old to the new partition layout, additional steps
are added to the platform upgrade hooks, to make the required changes
after OpenWrt has switch to the upgrade ramfs. The u-boot loader is
included as well, so it can be installed in the YAFFS during the
migration.

The migration is only executed if the version included in the metadata
of the sysupgrade image does not contain "23.05", so it is possible to
upgrade to older images until the migration is actually required.

Signed-off-by: Fabian Bläse <fabian@blaese.de>