Clean-up fdff handling
This commit addresses the issues with the fdff prefix by moving static
fdff addresses from the client interface to the loopback interface,
marking them as deprecated, and announcing the fdff::/64 prefix to
clients only when no other prefixes are configured.
This maintains the convenience of a short static IP while solving most
of the associated problems caused by using fdff addresses as
source addresses for external communication by the router and clients.

The following addresses will be moved to `lo`:

- fdff::1/128
- fdff::$MAC/128
- fdff::$EUI64/128

Additionally remove the unused `ula_prefix` option.

Fixes: #76

Signed-off-by: Johannes Kimmel <fff@bareminimum.eu>