fff-layer3-config: use dedicated netifd interface config for routerip
Previously, all routerip addresses were added to the loopback interface
config. To be able to clean up old addresses removed by the user, all
addresses are deleted from the loopback config. This also removes the
localhost addresses (127.0.0.1, ::1) from the interface configuration.

Under some circumstances, this can lead to the removal of the localhost
addresses from the kernel interface.

Instead, use a dedicated 'routerip' interface configuration, and leave
the loopback interface configuration as is.

Fixes: #340

Signed-off-by: Fabian Bläse <fabian@blaese.de>