Select ath10k by replacing ath10k-ct in the default packages of all targets
An issue with the OpenWrt build system leads to kernal symbols necessary
for the ath10k driver to get stripped. This is caused due to both ath10k
and ath10-ct generating kernel modules with identical names.

Work around this problem by replacing the targets default packages, so
ath10k-ct is not build as a module.

https://github.com/openwrt/openwrt/issues/14749

Fixes: #343

Signed-off-by: Fabian Bläse <fabian@blaese.de>