nodewatcher: remove alfred dependency
Up until now, the nodewatcher posted its result directly into alfred,
resuling in a dependency to alfred, and a rather complicated detour if
alfred is not the desired destination for the data (e.g. in the layer3
variant).

Instead, remove the cronjob for nodewatcher entirely, and let
nodewatcher output its data on stdout. New wrappers can then be added to
execute nodewatcher and post the resulting data to the appropriate
destination.

Fixes: #120

Signed-off-by: Fabian Bläse <fabian@blaese.de>