fff-network: move static client interface configuration to fff-node
In the layer3 variant, all client interfaces are configured in the
gateway configuration. Previously, the layer3 variant utilized the
hardcoded client interface for initial configuration, but since a
default gateway configuration is now generated on first boot, the
hardcoded client interface is now only required for the node variant.

Remove the hardcoded interface from the shared config and move it to
the node-spcific network configuration.

Signed-off-by: Fabian Bläse <fabian@blaese.de>