mt7621: backport edgerouter-x partition layout change
Due to increased kernel sizes with newer releases, the two kernel
partitions of the edgerouter-x and edgerouter-x-sfp have been merged
upstream with OpenWrt 24.10. Although the compat_version has been bumped
upstream, up- and downgrades are actually still possible, as long as the
kernel is small enough to fit into one of the two kernel partitions.

A mismatch of the kernels actual location and the embedded mtd
partitions is not a problem, because the kernel is loaded by the
bootloader, and the kernel partition is not used by OpenWrt itself.
Except for the installation upgrades, of course, but OpenWrt does
configure the bootloader accordingly.

To make upgrades easier, backport the partition layout change to our
firmware before upgrading to OpenWrt 24.10. Keep the old compat_version
for the upgrade files, so the installation is possible without a
warning. Bump it after successful installation, so future upgrade have
the same compat_version as upstream.

Signed-off-by: Fabian Bläse <fabian@blaese.de>