mvebu: rb5009: fix partition layout migration
While the partition layout migration does work, the u-boot included in
the firmware got mangled by strip/patchelf, which leads to missing or
corrupt data inside the bootloader.

Rebooting works, but powering the device of for a longer period of time
leads to boot failure.

Use gzip to prevent the OpenWrt build system to apply strip/patchelf to
the u-boot loader.

While at it, use the correct name for the new YAFFS partition.

Fixes: 808b0c7adc9b93beafa044258e5a27d83ea1e6d3

Signed-off-by: Fabian Bläse <fabian@blaese.de>