netifd: add pending patch to restore disable_ipv6 sysctl
If a device is added to a bond or bridge, the disable_ipv6 sysctl is
set, so that no link local addresses are present on the slave devices.

However, the disable_ipv6 sysctl is not restored after removing the
device from the bridge or bond. Therefore, no IPv6 link local addresses
are added and autoconfiguration does not work.

Set the disable_ipv6 to '0' on device removal from bridge or bond.

Fixes: #272
Fixes: dc48c9c3be8e7c8824e3680aefa4c61d10769699
References: https://github.com/openwrt/netifd/pull/40

Signed-off-by: Fabian Bläse <fabian@blaese.de>