ramips: fix accidental compat_version bump
With the previous firmware release, the compat_version for the Edgerouter
X has been incremented, due to the partition layout change. However, the
compat_version has mistakenly not only been incremented for the
Edgerouter, but for all devices of the ramips target.

Bump the compat_version of the upgrade image, so the next firmware
release can be installed without an error message - at least if the
updates are installed one by one.

This change is supposed to be reverted after the next release, because
the runtime compat_version has not been changed. This is done to keep
the compat_version in sync with upstream, despite the accidental bump.

Signed-off-by: Fabian Bläse <fabian@blaese.de>