fff-network: use correct cpu port for Mikrotik RB5009
When adding support for the Mikrotik RB5009, the cpu port definition was
missed. Therefore, it falls back to the default, which is "0t". However,
the device does not use swconfig, so it should not have any cpu port.

Adding the additional non-existing port does not seem to break bridge
setup, so no behavioral change is expected.

Add the device to the cpu port definition to remove the incorrectly
assinged cpu port.

Fixes: 774eeade044f65b1e0c3de189baff42ac683882d

Signed-off-by: Fabian Bläse <fabian@blaese.de>