buildscript: postpone variant selection until build is started
Up until now, the selection of the primary variant package has been done
in the prebuild step. However, the early selection of the primary
variant leads to the inclusion of many variant-specific packages in the
`./builscript config openwrt` step.

This is unintentional, as the base OpenWrt configuration shall only
include common packages and device specific option. Move the variant
selection to the build step instead.

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Tested-by: Christian Dresel <freifunk@dresel.systems>