fff-node: fixup swconfig cpu port
The cpuport has been removed from our switchport definitions, to improve
layer3 compatibility. As a result, the cpuport has to be added to the
port list when configuring vlans.

While steps have been taken to add the cpuport, they are at the wrong
place. Move them to the correct file/step instead.

Fixes: d1fbed9a09e7193999e49042dc7d4019105e3000

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Tested-by: Christian Dresel <freifunk@dresel.systems>