fff-node: remove WAN configuration from shared configuration
In a previous change [1], most of the shared network configuration has been
moved to the fff-node package, because it is not required for the layer3
variant anymore. The configuration of the WAN vlan and ifname was not
moved, because at the time layer3 did still rely on it.

fff-layer3-config is now able to handle WAN configuration itself. For
new setups, the default configuration is created to contain the WAN
interface, and for exising setups the configuration is migrated.
Therefore, move WAN configuration to fff-node.

[1] 99f99d0a75999ac34928eab9901d0f4e0442ebe3

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Tested-by: Christian Dresel <freifunk@dresel.systems>