bird: expose babel interface "limit" option
By default bird considers neighbors to be up when 12 or more HELLO
packets within a window of 16 are received.

This poses an issue on routers with a single uplink that suffers from
either frequent intermittend connection losses or firmware issues that
drops multicast packets.

Bird (or babel in general) will consider these links to be down
frequently, causing routing updates to the whole network and a complete
outage for the router behind the problematic link for a longer time than
necessary.

The wireless type can help in these situation, but there is no real
benefit from the constantly changing costs.

Therefore expose the limit option to the user, so a very low value may
be set on problematic links.

Signed-off-by: Johannes Kimmel <fff@bareminimum.eu>
Reviewed-by: Fabian Bläse <fabian@blaese.de>