fff-layer3-config: add per client network dns
Add possibility to define own DNS-Server per client network with setting
up in the 'config client' section(s) in /etc/gateway/config
	list dns '<IPv6 Nameserver>'
	list dns '<IPv4 Nameserver>'

Without this setting it will still be used the dnsmasq with its upstream
DNS Server.

With this setting the client get the DNS-Server directly via DHCP and
the dnsmasq from firmware will not be used!

Possible workflows are for own DNS-Server, AdGuard and PI-Hole.

Closes: #381

Reviewed-by: Fabian Bläse <fabian@blaese.de>